requirements.txt
aioredis==2.0.1
aiosmtplib==2.0.2
alabaster==0.7.16
alembic==1.13.3
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
babel==2.16.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
colorama==0.4.6
commonmark==0.9.1
dnspython==2.6.1
docutils==0.20.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.73.0
fastapi-mail==1.4.1
fastapi-ratelimiter==0.0.2
greenlet==3.1.1
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.9
pyasn1==0.6.1
pydantic==1.10.18
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
recommonmark==0.7.1
requests==2.32.3
rsa==4.9
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.35
starlette==0.17.1
typing_extensions==4.12.2
urllib3==2.2.3
